Dosyalar yeniden düzenlendi ve değişiklikler yapıldı.
