--- kds proje site/iklim_hava_kosullari.html
@@ -1,157 +1,0 @@
-<!DOCTYPE html>
-<html lang="tr">
-
-<head>
-    <meta charset="UTF-8">
-
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-    <meta name="author" content="Ahmet Zafer Demir">
-
-    <link rel="stylesheet" type="text/css" href="style.css">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Montserrat">
-
-    <style>
-        * {
-    margin: 0px;
-    padding: 0px;
-    box-sizing: border-box;
-}
-
-html {
-    background-color: black;
-}
-
-.container {
-    width: 1500px;
-    min-height: 1000px;
-    background-color: #4c4c4c;
-    margin: auto;
-}
-
-nav {
-    width: 300px;
-    height: 1000px;
-    float: left;
-    background-color: black;
-    font-family: Montserrat, sans-serif, serif;
-}
-
-.profil-resmi {
-    position: relative;
-    top: 50px;
-    left: 100px;
-}
-
-.profil-resmi img {
-    width: 100px;
-    border-radius: 50%;
-    z-index: 0;
-}
-
-.ad-soyad-bilgisi {
-    width: 75%;
-    height: 75px;
-    position: relative;
-    top: 70px;
-    margin: auto;
-    color: white;
-    text-align: center;
-    font-family: Montserrat, sans-serif, serif;
-    text-transform: uppercase;
-    font-weight: bold;
-}
-
-.linkler {
-    margin-top: 75px;
-    width: 100%;
-    height: 400px;
-}
-
-.linkler li {
-    border-top: purple solid 1px;
-    border-bottom: purple solid 1px;
-    margin-top: 15px;
-    text-align: center;
-}
-
-.linkler li a {
-    display: block;
-    text-decoration: none;
-    color: white;
-    padding: 20px;
-}
-
-.linkler li a:hover {
-    font-weight: bold;
-    color: purple;
-    background-color: white;
-}
-
-.linkler li a:active {
-    font-weight: bold;
-    color: red;
-    background-color: white;
-}
-
-article {
-    float: left;
-    width: 1200px;
-    height: 1000px;
-    border: aqua solid 1px;
-    background-color: gray;
-    overflow: auto;
-}
-
-#urun-toprak-sayfasi-icerik {
-    padding: 50px;
-    font-family: "Montserrat",sans-serif,serif;
-}
-
-#urun-toprak-sayfasi-icerik h1{
-    color: purple;
-    font-size: 50px;
-}
-    </style>
-
-    <title>Tarla Yönetimi Portalı</title>
-</head>
-
-<body>
-
-    <div class="container">
-
-
-        <nav>
-            <div class="profil">
-                <div class="profil-resmi">
-                    <img src="../img/ben.jpg" alt="Ahmet Zafer Demir" title="Ahmet Zafer Demir">
-                </div>
-                <div class="ad-soyad-bilgisi">
-                    <span>Ahmet Zafer Demir</span>
-                </div>
-
-                <div class="linkler">
-                    <ul>
-                        <li><a href="iklim_hava_kosullari">İklim / Hava Koşulları</a></li>
-                        <li><a href="urun_toprak_bilgileri">Ürün / Toprak Tipi Bilgileri</a></li>
-                        <li><a href="grafikler">Grafikler</a></li>
-                        <li><a href="uretim_satis_verileri">Üretim / Satış Verileri</a></li>
-                        <li><a href="veri_girisi">Veri Girişi</a></li>
-                    </ul>
-                </div>
-            </div>
-        </nav>
-
-        <article class="icerik" id="urun-toprak-sayfasi-icerik">
-            <h1>İklim Hava Koşulları</h1>
-
-
-        </article>
-
-    </div>
-
-
-</body>
-
-</html>
