--- kds proje site/routers/grafikler_router.js
@@ -1,6 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-// ... Router'a yönlendirmeleri ekle ...
-
-module.exports = router;
