+++ routers/urun_toprak_router.js
@@ -1,0 +1,10 @@
+const express = require('express');
+const router = express.Router();
+
+// ... Router'a yönlendirmeleri ekle ...
+
+
+
+
+
+module.exports = router;
