+++ routers/veri_girisi_router.js
@@ -1,0 +1,6 @@
+const express = require('express');
+const router = express.Router();
+
+// ... Router'a yönlendirmeleri ekle ...
+
+module.exports = router;
