+++ veri_girisi.html
@@ -1,0 +1,278 @@
+<!DOCTYPE html>
+<html lang="tr">
+
+<head>
+    <meta charset="UTF-8">
+
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+    <meta name="author" content="Ahmet Zafer Demir">
+
+    <link rel="stylesheet" type="text/css" href="style.css">
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Montserrat">
+
+    <style>
+        * {
+            margin: 0px;
+            padding: 0px;
+            box-sizing: border-box;
+        }
+
+        html {
+            background-color: black;
+        }
+
+        .container {
+            width: 1500px;
+            min-height: 1000px;
+            background-color: #4c4c4c;
+            margin: auto;
+        }
+
+        nav {
+            width: 300px;
+            height: 1000px;
+            float: left;
+            background-color: black;
+            font-family: Montserrat, sans-serif, serif;
+        }
+
+        .profil-resmi {
+            position: relative;
+            top: 50px;
+            left: 100px;
+        }
+
+        .profil-resmi img {
+            width: 100px;
+            border-radius: 50%;
+            z-index: 0;
+        }
+
+        .ad-soyad-bilgisi {
+            width: 75%;
+            height: 75px;
+            position: relative;
+            top: 70px;
+            margin: auto;
+            color: white;
+            text-align: center;
+            font-family: Montserrat, sans-serif, serif;
+            text-transform: uppercase;
+            font-weight: bold;
+        }
+
+        .linkler {
+            margin-top: 75px;
+            width: 100%;
+            height: 400px;
+        }
+
+        .linkler li {
+            border-top: purple solid 1px;
+            border-bottom: purple solid 1px;
+            margin-top: 15px;
+            text-align: center;
+        }
+
+        .linkler li a {
+            display: block;
+            text-decoration: none;
+            color: white;
+            padding: 20px;
+        }
+
+        .linkler li a:hover {
+            font-weight: bold;
+            color: purple;
+            background-color: white;
+        }
+
+        .linkler li a:active {
+            font-weight: bold;
+            color: red;
+            background-color: white;
+        }
+
+        article {
+            float: left;
+            width: 1200px;
+            height: 1000px;
+            border: aqua solid 1px;
+            background-color: gray;
+            overflow: auto;
+        }
+
+        #urun-toprak-sayfasi-icerik {
+            padding: 50px;
+            font-family: "Montserrat", sans-serif, serif;
+        }
+
+        #urun-toprak-sayfasi-icerik h1 {
+            color: purple;
+            font-size: 50px;
+        }
+
+        h2 {
+            margin: 40px 0px 20px 10px;
+        }
+
+        .genel_div1 {
+            margin-left: 20px;
+            width: 40%;
+        }
+
+        label {
+            color: white;
+            font-weight: bold;
+            display: block;
+        }
+
+        input {
+            border-left-width: 10px;
+            border-left-color: purple;
+            border-radius: 5px;
+            padding: 5px;
+            font-weight: bolder;
+        }
+
+        .tarla_giris_item {
+            display: block;
+            margin-top: 10px;
+        }
+
+        button {
+            padding: 5px;
+            margin: 30px 0px 0px 0px;
+            font-weight: bold
+        }
+
+        button:hover {
+            color: purple;
+        }
+
+        .genel_div2 {
+            border: solid 1px purple;
+            width: 100%;
+            margin-top: 25px;
+        }
+
+        .genel_div3 {
+            margin-left: 20px;
+            width: 40%;
+        }
+
+        .urun_giris_item {
+            display: block;
+            margin-top: 10px;
+        }
+
+        .genel_div4 {
+            border: solid 1px purple;
+            width: 100%;
+            margin-top: 25px;
+
+        }
+    </style>
+
+    <title>Tarla Yönetimi Portalı</title>
+</head>
+
+<body>
+
+    <div class="container">
+
+
+        <nav>
+            <div class="profil">
+                <div class="profil-resmi">
+                    <img src="../img/ben.jpg" alt="Ahmet Zafer Demir" title="Ahmet Zafer Demir">
+                </div>
+                <div class="ad-soyad-bilgisi">
+                    <span>Ahmet Zafer Demir</span>
+                </div>
+
+                <div class="linkler">
+                    <ul>
+                        <li><a href="iklim_hava_kosullari">İklim / Hava Koşulları</a></li>
+                        <li><a href="urun_toprak_bilgileri">Ürün / Toprak Tipi Bilgileri</a></li>
+                        <li><a href="grafikler">Grafikler</a></li>
+                        <li><a href="uretim_satis_verileri">Üretim / Satış Verileri</a></li>
+                        <li><a href="veri_girisi">Veri Girişi</a></li>
+                    </ul>
+                </div>
+            </div>
+        </nav>
+
+        <article class="icerik" id="urun-toprak-sayfasi-icerik">
+            <h1>Veri Girişi</h1>
+
+            <h2>Yeni Tarla Girişi</h2>
+            <div class="genel_div1">
+                <div class="tarla_veri_giris_div">
+                    <div class="tarla_id tarla_giris_item">
+                        <label for="tarla_id">Tarla ID</label>
+                        <input type="number" name="tarla_id" id="tarla_id">
+                    </div>
+
+                    <div class="urun_id tarla_giris_item">
+                        <label for="urun_id">Ürün ID</label>
+                        <input type="number" name="urun_id" id="urun_id">
+                    </div>
+
+                    <div class="tarla_m2 tarla_giris_item">
+                        <label for="tarla_m2">Tarla m2</label>
+                        <input type="number" name="tarla_m2" id="tarla_m2">
+                    </div>
+
+                    <div class="toprak_tip_id tarla_giris_item">
+                        <label for="toprak_tip_id">Toprak Tipi ID</label>
+                        <input type="number" name="toprak_tip_id" id="toprak_tip_id">
+                    </div>
+                </div>
+
+                <button id="tarla_ekle">Ekle</button>
+
+            </div>
+
+
+
+            <div class="genel_div2">
+                Tarla Tablosu Gelecek
+            </div>
+
+
+
+            <h2>Yeni Ürün Girişi</h2>
+            <div class="genel_div3">
+                <div class="urun_veri_giris_div">
+                    <div class="urun_id urun_giris_item">
+                        <label for="urun_id">Ürün ID</label>
+                        <input type="number" name="urun_id" id="urun_id">
+                    </div>
+
+                    <div class="urun_ad urun_giris_item">
+                        <label for="urun_ad">Ürün Ad</label>
+                        <input type="text" name="urun_ad" id="urun_ad">
+                    </div>
+
+                </div>
+
+                <button id="urun_ekle">Ekle</button>
+
+            </div>
+
+
+            <div class="genel_div4">
+                Ürün Tablosu Gelecek
+            </div>
+
+            
+        </article>
+
+    </div>
+
+
+</body>
+
+</html>
